Add basic integration test, refactor package structure
