+++ src/main/java/de/nloewes/roshambr/RoshambrApplication.java
@@ -2,8 +2,9 @@ package de.nloewes.roshambr;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class}) //TODO
 public class RoshambrApplication {
 
 	public static void main(String[] args) {
